// cluster/console/console/web/package/src/apis/cordiumv1/cordiumv1.ts
export interface Workspace_Spec_Repository {
    /**
     * URL is the HTTPS URL of the git repository (e.g.
     * `https://github.com/myorg/my-project`). Only the `https` scheme is
     * supported.
     *
     * @generated from protobuf field: string url = 1
     */
    url: string;
    /**
     * CloneOptions controls how the repository is cloned.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Repository.CloneOptions cloneOptions = 2
     */
    cloneOptions?: Workspace_Spec_Repository_CloneOptions;
    /**
     * Authentication is set for the private repositories that require
     * authentication.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Repository.Authentication authentication = 3
     */
    authentication?: Workspace_Spec_Repository_Authentication;
}
// ...
export interface Template {
    /**
     * APIVersion is the API version (i.e. "cordium/v1")
     *
     * @generated from protobuf field: string apiVersion = 1
     */
    apiVersion: string;
    /**
     * Kind is the resource name (i.e. `Template`).
     *
     * @generated from protobuf field: string kind = 2
     */
    kind: string;
    /**
     * Metadata is the object's metadata.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.Metadata metadata = 3
     */
    metadata?: Metadata;
    /**
     * Spec is the Template specification.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Template.Spec spec = 4
     */
    spec?: Template_Spec;
    /**
     * Status is the current status of the Template.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Template.Status status = 5
     */
    status?: Template_Status;
}
/**
 * Spec is the Template specification. It shares most of the Workspace spec.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Template.Spec
 */
export interface Template_Spec {
    /**
     * Image defines how the container image of the Template's Workspaces is
     * obtained.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Image image = 1
     */
    image?: Workspace_Spec_Image;
    /**
     * Runtime controls the behavior of the container of the Template's
     * Workspaces.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Runtime runtime = 2
     */
    runtime?: Workspace_Spec_Runtime;
    /**
     * Repository is the primary git repository which is cloned into the
     * Template's Workspaces.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Repository repository = 3
     */
    repository?: Workspace_Spec_Repository;
    /**
     * AdditionalRepositories is the list of the secondary repositories that
     * are cloned alongside the primary one.
     *
     * @generated from protobuf field: repeated octelium.api.main.cordium.v1.Workspace.Spec.AdditionalRepository additionalRepositories = 4
     */
    additionalRepositories: Workspace_Spec_AdditionalRepository[];
    /**
     * Limit is the default compute resources that are allocated for the
     * Template's Workspaces.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec.Limit limit = 5
     */
    limit?: Workspace_Spec_Limit;
    /**
     * GitProvider is the name of a GitProvider in the same Space. Once it is
     * set, the User's stored OAuth2 token is automatically injected into the
     * Template's Workspaces which enables authenticated git operations without
     * any manual credential configuration.
     *
     * @generated from protobuf field: string gitProvider = 6
     */
    gitProvider: string;
    /**
     * Vars is the list of the variables that are substituted inside the spec.
     *
     * @generated from protobuf field: repeated octelium.api.main.cordium.v1.Workspace.Spec.Var vars = 7
     */
    vars: Workspace_Spec_Var[];
}
/**
 * Status is the current status of the Template. It is managed by the Cluster
 * and it is read-only.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.Template.Status
 */
export interface Template_Status {
    /**
     * SpaceRef is the reference of the Space that owns the Template.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference spaceRef = 1
     */
    spaceRef?: ObjectReference;
    /**
     * UserRef is the reference of the Octelium User who created the Template.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference userRef = 2
     */
    userRef?: ObjectReference;
    /**
     * GitProviderRef is the reference of the GitProvider that is associated
     * with the Template via the spec's gitProvider field.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference gitProviderRef = 3
     */
    gitProviderRef?: ObjectReference;
    /**
     * BuildInfo is the state of the Template's pre-builds.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Template.Status.BuildInfo buildInfo = 4
     */
    buildInfo?: Template_Status_BuildInfo;
}
// ...
export interface Template_Status_BuildInfo_Build {
    /**
     * ID is a randomly generated identifier of the pre-build.
     *
     * @generated from protobuf field: string id = 1
     */
    id: string;
    /**
     * Tags is the list of the tags that the pre-build was started with. It
     * defaults to `latest`.
     *
     * @generated from protobuf field: repeated string tags = 2
     */
    tags: string[];
    /**
     * StartedAt is the timestamp at which the pre-build started.
     *
     * @generated from protobuf field: google.protobuf.Timestamp startedAt = 3
     */
    startedAt?: Timestamp;
    /**
     * DoneAt is the timestamp at which the pre-build completed, failed or
     * was canceled.
     *
     * @generated from protobuf field: google.protobuf.Timestamp doneAt = 4
     */
    doneAt?: Timestamp;
    /**
     * IsCanceled means that the pre-build was explicitly canceled (e.g.
     * via the CancelBuildTemplate method or by starting a new pre-build).
     *
     * @generated from protobuf field: bool isCanceled = 5
     */
    isCanceled: boolean;
    /**
     * Failure is the reason of the failure of the pre-build, if any.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Status.Failure failure = 6
     */
    failure?: Workspace_Status_Failure;
    /**
     * State is the current state of the pre-build.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Template.Status.BuildInfo.Build.State state = 7
     */
    state: Template_Status_BuildInfo_Build_State;
}
/**
 * State is the current state of the pre-build
 *
 * @generated from protobuf enum octelium.api.main.cordium.v1.Template.Status.BuildInfo.Build.State
 */
export enum Template_Status_BuildInfo_Build_State {
    /**
     * STATE_UNKNOWN is not used.
     *
     * @generated from protobuf enum value: STATE_UNKNOWN = 0;
     */
    UNKNOWN = 0,
    /**
     * STATE_RUNNING means that the pre-build is currently running.
     *
     * @generated from protobuf enum value: STATE_RUNNING = 1;
     */
    RUNNING = 1,
    /**
     * STATE_READY means that the pre-build successfully completed and
     * that its storage snapshot can be used by the new Workspaces.
     *
     * @generated from protobuf enum value: STATE_READY = 2;
     */
    READY = 2,
    /**
     * STATE_FAILED means that the pre-build failed or that it was
     * canceled.
     *
     * @generated from protobuf enum value: STATE_FAILED = 3;
     */
    FAILED = 3
}
// ...
export const Workspace_Spec_Image = new Workspace_Spec_Image$Type();
// ...
export const Workspace_Spec_Repository = new Workspace_Spec_Repository$Type();
// ...
export const Workspace_Spec_AdditionalRepository = new Workspace_Spec_AdditionalRepository$Type();
// ...
export const Workspace_Spec_Runtime = new Workspace_Spec_Runtime$Type();
// ...
export const Workspace_Spec_Limit = new Workspace_Spec_Limit$Type();
// ...
export const Workspace_Spec_Var = new Workspace_Spec_Var$Type();
// ...
export const Workspace_Status_Failure = new Workspace_Status_Failure$Type();
// ...
class Template$Type extends MessageType<Template> {
    constructor() {
        super("octelium.api.main.cordium.v1.Template", [
            { no: 1, name: "apiVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "kind", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "metadata", kind: "message", T: () => Metadata },
            { no: 4, name: "spec", kind: "message", T: () => Template_Spec },
            { no: 5, name: "status", kind: "message", T: () => Template_Status }
        ]);
    }
    create(value?: PartialMessage<Template>): Template {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.apiVersion = "";
        message.kind = "";
        if (value !== undefined)
            reflectionMergePartial<Template>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Template): Template {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string apiVersion */ 1:
                    message.apiVersion = reader.string();
                    break;
                case /* string kind */ 2:
                    message.kind = reader.string();
                    break;
                case /* octelium.api.main.meta.v1.Metadata metadata */ 3:
                    message.metadata = Metadata.internalBinaryRead(reader, reader.uint32(), options, message.metadata);
                    break;
                case /* octelium.api.main.cordium.v1.Template.Spec spec */ 4:
                    message.spec = Template_Spec.internalBinaryRead(reader, reader.uint32(), options, message.spec);
                    break;
                case /* octelium.api.main.cordium.v1.Template.Status status */ 5:
                    message.status = Template_Status.internalBinaryRead(reader, reader.uint32(), options, message.status);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Template, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string apiVersion = 1; */
        if (message.apiVersion !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.apiVersion);
        /* string kind = 2; */
        if (message.kind !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.kind);
        /* octelium.api.main.meta.v1.Metadata metadata = 3; */
        if (message.metadata)
            Metadata.internalBinaryWrite(message.metadata, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.cordium.v1.Template.Spec spec = 4; */
        if (message.spec)
            Template_Spec.internalBinaryWrite(message.spec, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.cordium.v1.Template.Status status = 5; */
        if (message.status)
            Template_Status.internalBinaryWrite(message.status, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.cordium.v1.Template
 */
export const Template = new Template$Type();
// ...
export const Template_Spec = new Template_Spec$Type();
// ...
export const Template_Status = new Template_Status$Type();
// ...
export const Template_Status_BuildInfo = new Template_Status_BuildInfo$Type();
// ...
export const Template_Status_BuildInfo_Build = new Template_Status_BuildInfo_Build$Type();